requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==2.21.0
dill==0.3.8
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fasttext==0.9.3
fasttext-numpy2==0.10.4
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
mteb @ file:///Users/cpondoc/Desktop/research/noisy-benchmark/mteb
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
openai==1.63.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
polars==1.22.0
propcache==0.3.0
pyarrow==19.0.1
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrec_eval-terrier==0.5.6
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
